pyproject.toml
[project]
name = "disaster-response"
version = "1.0.0"
description = "DisasterResponseEnv — Emergency Operations Center Triage Agent for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "models.py", "inference.py"]

[tool.uv.sources]

[tool.pytest.ini_options]
testpaths = ["tests"]
